// frontend/src/api/catalogService/model/product.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProductCategory } from './productCategory';

export interface Product {
  id?: string;
  name?: string;
  description?: string;
  price?: number;
  category?: ProductCategory;
  imageUrl?: string;
  ingredients?: string[];
  available?: boolean;
}