// frontend/src/api/catalogService/model/productCategory.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductCategory = {
  PIZZA: 'PIZZA',
  DRINK: 'DRINK',
  DESSERT: 'DESSERT',
  SIDE: 'SIDE',
  SAUCE: 'SAUCE',
} as const;