// frontend/src/api/catalogService/model/productRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProductRequestCategory } from './productRequestCategory';

export interface ProductRequest {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  category: ProductRequestCategory;
  imageUrl?: string;
  ingredients?: string[];
  available?: boolean;
}